backend/requirements.txt
# Agent: Kiran | Sprint: 01 | Date: 2026-03-16
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
anthropic>=0.40.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.27.0
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-cov==5.0.0
anyio==4.3.0
python-dotenv==1.0.1
aiosqlite==0.20.0
